cooja: use C99 uintptr_t

+++ arch/platform/cooja/sys/cooja_mtarch.h
@@ -35,13 +35,15 @@
 #ifndef COOJA_MTARCH_H_
 #define COOJA_MTARCH_H_
 
+#include <stdint.h>
+
 #ifndef COOJA_MTARCH_STACKSIZE
 #define COOJA_MTARCH_STACKSIZE 1024
 #endif /* COOJA_MTARCH_STACKSIZE */
 
 struct cooja_mtarch_thread {
-  unsigned long sp;  /* Note: stack pointer must be first var in struct! */
-  unsigned long stack[COOJA_MTARCH_STACKSIZE];
+  uintptr_t sp;  /* Note: stack pointer must be first var in struct! */
+  uintptr_t stack[COOJA_MTARCH_STACKSIZE];
 } __attribute__ ((aligned (16)));
 
 struct cooja_mt_thread;
